+++ CHANGELOG.md
@@ -20,12 +20,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Closing HTML/XML tags works inconsistently in Notepad++ v7.3.3.
 - Avoid plugin crash when too-large int values are entered in the selection-remembering form.
+- Holding down `Enter` in a multiline textbox does not add multiple new lines; it only adds one newline on keyup.
 
 ## [0.0.3] - (UNRELEASED) YYYY-MM-DD
 
 ### Fixed
 
 1. Greatly reduced unnecessary calls to `NPPM_GETCURRENTLANGTYPE` in `DoInsertHtmlCloseTag` by caching the lexer language whenever the lexer language changes or a buffer is opened.
+2. Bug where hitting `Enter` in a multiline textbox would not add a new line.
 
 ## [0.0.2] - 2024-02-06